Blocks/HTTP/request: Added `url_original` to `Request`.
Will need it to check if the original URL does not end with a slash. The request handler `Register` requires the routes to not end with a slash, so there is no way to register two separate handlers for a URL with and without trailing slash, this distinction has to be made inside the handler.