ServeStaticFilesFrom with separate route and URL prefixes. URL redirects refactor.
The issue I'm solving here is issuing a redirect from the backend with the URL understood by the end user, even if the URL is transformed along the way from the server to the user (e.g. on a proxy).

Unfortunately, this required the changes to the `URL` class behavior: its "previous URL" handling and defaults handling. There were too many variants of use of this class's options, some of which were conflicting with each other (e.g. default scheme with an unknown port or no scheme with an unknown port).

The `HTTPClient` was also changed to throw if the redirects weren't allowed before following any redirect response code.

The `SocketResolveAddressException` was filled with the resolve arguments to have more details about the error in runtime.

CC @dkorolev @grixa @mzhurovich