TypeSystem/Schema: For autogenerated TypeScript schemas, fix Variant types.
Previously, Variant types were encoded as unions of the respective types, but in the JSON serialized structure the Variant case is an object with a key equal to the type name, and an empty-string key with the type identifier.

I had to restructure the entire TypeScript schema generator to support that.